Add features from the "further updates" list

+++ backend/src/main/java/com/article/backend/controller/ImageController.java
@@ -1,18 +1,15 @@
 package com.article.backend.controller;
 
-import com.article.backend.exception.FileStorageException;
 import com.article.backend.model.Article;
 import com.article.backend.model.ArticleImage;
 import com.article.backend.service.ArticleImageService;
 import com.article.backend.service.ArticleService;
+import com.article.backend.util.FileToBase64Utils;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.util.Base64;
 
 @RestController
 @RequestMapping("/image")
@@ -29,22 +26,21 @@ public class ImageController {
                                                   @RequestPart MultipartFile file) {
         Article article = articleService.getArticleById(articleId);
         if (article == null) {
-           throw new EntityNotFoundException("Article not found.");
+            throw new EntityNotFoundException("Article not found.");
         }
-
-        try {
-            ArticleImage newArticleImage = new ArticleImage();
-            newArticleImage.setArticle(article);
-            newArticleImage.setName(file.getOriginalFilename());
-            newArticleImage.setImage(Base64.getEncoder().encodeToString(file.getBytes()));
-            return ResponseEntity.ok(articleImageService.saveArticleImage(newArticleImage));
-        } catch (IOException e) {
-            throw new FileStorageException("Could not store the file. Please try again!");
-        }
+        ArticleImage newArticleImage = new ArticleImage();
+        newArticleImage.setArticle(article);
+        newArticleImage.setName(file.getOriginalFilename());
+        newArticleImage.setImage(FileToBase64Utils.getBase64ImageFromFile(file));
+        return ResponseEntity.ok(articleImageService.saveArticleImage(newArticleImage));
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deleteImage(@RequestParam Long id) {
+        ArticleImage existingArticleImage = articleImageService.getArticleImageById(id);
+        if (existingArticleImage == null) {
+            throw new EntityNotFoundException("ArticleImage not found.");
+        }
         articleImageService.deleteById(id);
         return ResponseEntity.ok().build();
     }
